--- a/src/Acquisition.jl
+++ b/src/Acquisition.jl
@@ -1,9 +1,24 @@
 module Acquisition
 
-    using DocStringExtensions, GNSSSignals, RecipesBase, FFTW, Statistics, LinearAlgebra, LoopVectorization, Unitful
+    using DocStringExtensions,
+        GNSSSignals,
+        RecipesBase,
+        FFTW,
+        Statistics,
+        LinearAlgebra,
+        LoopVectorization,
+        Unitful
+
     import Unitful: s, Hz
 
-    export acquire, plot_acquisition_results, coarse_fine_acquire
+    export
+        acquire,
+        plot_acquisition_results,
+        coarse_fine_acquire,
+        coarse_fine_acquire!,
+        acquire!,
+        AcquisitionPlan,
+        CoarseFineAcquisitionPlan
 
     struct AcquisitionResults{S<:AbstractGNSS, T}
         system::S
@@ -12,10 +27,12 @@ module Acquisition
         carrier_doppler::typeof(1.0Hz)
         code_phase::Float64
         CN0::Float64
-        power_bins::Array{T, 2}
+        noise_power::T
+        power_bins::Matrix{T}
         dopplers::StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}
     end
 
+    include("plan_acquire.jl")
     include("downconvert.jl")
     include("plot.jl")
     include("calc_powers.jl")