--- a/test/acquire.jl
+++ b/test/acquire.jl
@@ -1,5 +1,5 @@
 @testset "Acquire signal $(get_system_string(system)) and signal type $type" for system in [GPSL1(), GalileoE1B()], type in [Float64, Float32, Int16, Int32]
-    Random.seed!(1234)
+    Random.seed!(2345)
     num_samples = 20000
     doppler = 1234Hz
     code_phase = 110.613261