This pull request increases the performance:
Before:
```julia
using Acquisition, BenchmarkTools, GNSSSignals
using Unitful: Hz
signal = rand(Complex{Int16}, 20000)
gpsl1 = GPSL1()
@btime [coarse_fine_acquire($gpsl1, $signal, $(5e6Hz), prn) for prn = 1:32]
# 11.033 s (110979 allocations: 27.57 GiB)
gal = GalileoE1B()
@btime [coarse_fine_acquire($gal, $signal, $(5e6Hz), prn) for prn = 1:32]
#  24.143 s (110979 allocations: 80.80 GiB)
```
After:
```julia
@btime coarse_fine_acquire($gpsl1, $signal, $(5e6Hz), $(1:32))
#  1.384 s (21520 allocations: 274.79 MiB)
@btime coarse_fine_acquire($gal, $signal, $(5e6Hz), $(1:32))
#  2.178 s (21515 allocations: 954.83 MiB)
```
The syntax changed a little bit. However, the old syntax is still valid. So this is not a breaking change.
The new syntax is required for the high performance implementation.