@zsoerenm 2022-11-22
I increased the performance even further.
```julia
@btime coarse_fine_acquire($gpsl1, $signal, $(5e6Hz), $(1:32))
#  401.770 ms (20013 allocations: 126.18 MiB)
@btime coarse_fine_acquire($gal, $signal, $(5e6Hz), $(1:32))
#  567.347 ms (20021 allocations: 466.75 MiB)
```
If you have to acquire satellites multiple times in a row, there is now the possibility to plan the acquisition upfront which ultimately increases the performance even further ;)
```julia
acq_plan = CoarseFineAcquisitionPlan(gpsl1, length(signal), 5e6Hz)
@btime coarse_fine_acquire!($acq_plan, $signal, $(1:32))
#  390.803 ms (19629 allocations: 908.88 KiB)
gal_acq_plan = CoarseFineAcquisitionPlan(gal, length(signal), 5e6Hz)
@btime coarse_fine_acquire!($gal_acq_plan, $signal, $(1:32))
#  491.790 ms (19637 allocations: 909.00 KiB)
```
Also note the significant drop in allocations. This might be handy for real time applications.
@dominformant Would you mind reviewing this as well? ;)