@dominformant 2022-11-22
Nice work! I would just prefer, if we could rename `coarse_fine_acquisition!` to just `acquisition!` in case that a plan is supplied. This would allow to switch between the acuisition types without the need to change the rest of the code. For the "plan-less" acquisition, the different function names are fine as they allow to distinguish between the methods.