@zsoerenm 2022-11-22
Yes, I thought about that, too. I will change that tomorrow.